chore: set all type style properties explicitly
Because CSS properties such as font-weight are inherited,
if a parent element sets a type style that sets a medium font weight
and a child sets a style that doesn't set font weight, but is expecting
normal weight, then the child will render with an unexpected medium
font weight.

The same goes for text-rendering, letter-spacing and font-family, so
all properties need to be explicitly set for all type styles.